Move all binding functions to ci_binding.go and unify names
