+++ infocmdb/ci_binding.go
@@ -1,10 +1,13 @@
 package infocmdb
 
 import (
+	"errors"
 	"fmt"
 	"reflect"
 	"strconv"
 	"strings"
+
+	v2 "github.com/infonova/infocmdb-sdk-go/infocmdb/v2/infocmdb"
 )
 
 type BindError struct {
@@ -19,7 +22,83 @@ func (e *BindError) Error() string {
 	return e.Msg
 }
 
-func bindCiAttributes(attributes []CiAttribute, out interface{}) (err error) {
+func (c *Client) GetAndBindCi(ciId int, out interface{}) (err error) {
+	attributes, err := c.GetCiAttributes(ciId)
+	if err != nil {
+		return
+	}
+
+	return bindCi(ciId, attributes, out)
+}
+
+func (c *Client) GetAndBindListOfCis(ciIds []int, out interface{}) (err error) {
+	ciIdToAttributesMap, err := c.GetMapOfCiAttributes(ciIds)
+	if err != nil {
+		return
+	}
+
+	outSlicePtr := reflect.ValueOf(out)
+	if outSlicePtr.Kind() != reflect.Ptr {
+		return errors.New("out parameter is not a slice pointer")
+	}
+	outSlice := outSlicePtr.Elem()
+	if outSlice.Kind() != reflect.Slice {
+		return errors.New("out parameter is not a slice pointer")
+	}
+	outSliceElem := reflect.TypeOf(outSlice.Interface()).Elem()
+	outSliceValue := reflect.MakeSlice(outSlice.Type(), 0, 0)
+
+	for _, ciId := range ciIds {
+		ciAttributes := ciIdToAttributesMap[ciId]
+
+		elem := reflect.New(outSliceElem)
+		err = bindCi(ciId, ciAttributes, elem.Interface())
+		if err != nil {
+			return
+		}
+
+		outSliceValue = reflect.Append(outSliceValue, elem.Elem())
+	}
+
+	outSlice.Set(outSliceValue)
+	return
+}
+
+func (c *Client) GetAndBindListOfCisOfCiTypeName(ciTypeName string, out interface{}) (err error) {
+	ciIds, err := c.GetListOfCiIdsOfCiTypeName(ciTypeName)
+	if err != nil {
+		err = errors.New("failed to get \"" + ciTypeName + "\" ci ids: " + err.Error())
+		return
+	}
+
+	err = c.GetAndBindListOfCis(ciIds, out)
+	return
+}
+
+func (c *Client) GetAndBindCiByAttributeValue(name string, value string, valueType v2.AttributeValueType, out interface{}) (err error) {
+	ciId, err := c.GetCiIdByAttributeValue(name, value, valueType)
+	if err != nil {
+		return
+	}
+
+	err = c.GetAndBindCi(ciId, out)
+	if err != nil {
+		return
+	}
+
+	return
+}
+
+func (c *Client) GetAndBindCiByAttributeValueText(name string, value string, out interface{}) (err error) {
+	return c.GetAndBindCiByAttributeValue(name, value, v2.ATTRIBUTE_VALUE_TYPE_TEXT, out)
+}
+
+func (c *Client) GetAndBindCiByAttributeValueCi(name string, value string, out interface{}) (err error) {
+	return c.GetAndBindCiByAttributeValue(name, value, v2.ATTRIBUTE_VALUE_TYPE_CI, out)
+}
+
+
+func bindCi(ciId int, attributes []CiAttribute, out interface{}) (err error) {
 	attrNameToAttrMap := map[string][]CiAttribute{}
 	for _, attr := range attributes {
 		ciAttributes := attrNameToAttrMap[attr.AttributeName]
@@ -90,7 +169,8 @@ func bindAttrToStringField(attrs []CiAttribute, field reflect.Value) (err error)
 	} else {
 		attr := attrs[0]
 		return &BindError{
-			Msg:       fmt.Sprintf("failed to map multiple attributes with name \"%v\" to string", attr.AttributeName),
+			Msg: fmt.Sprintf("failed to map multiple attributes with name \"%v\" to string",
+				attr.AttributeName),
 			FieldName: field.Type().Name(),
 			SrcName:   attr.AttributeName,
 			SrcType:   attr.AttributeType,
@@ -108,7 +188,7 @@ func bindAttrToIntField(attrs []CiAttribute, field reflect.Value) (err error) {
 		intValue, err := strconv.Atoi(attr.Value)
 		if err != nil {
 			return &BindError{
-				Msg:       fmt.Sprintf("failed to map attribute with name \"%v\" and value \"%v\" to int",
+				Msg: fmt.Sprintf("failed to map attribute with name \"%v\" and value \"%v\" to int",
 					attr.AttributeName, attr.Value),
 				FieldName: field.Type().Name(),
 				SrcName:   attr.AttributeName,
@@ -119,7 +199,8 @@ func bindAttrToIntField(attrs []CiAttribute, field reflect.Value) (err error) {
 	} else {
 		attr := attrs[0]
 		return &BindError{
-			Msg:       fmt.Sprintf("failed to map multiple attributes with name \"%v\" to int", attr.AttributeName),
+			Msg: fmt.Sprintf("failed to map multiple attributes with name \"%v\" to int",
+				attr.AttributeName),
 			FieldName: field.Type().Name(),
 			SrcName:   attr.AttributeName,
 			SrcType:   attr.AttributeType,
@@ -147,7 +228,8 @@ func bindAttrToStringSliceField(attrs []CiAttribute, field reflect.Value) (err e
 			}
 		default:
 			return &BindError{
-				Msg:       fmt.Sprintf("failed to map attribute type %v to []string", attr.AttributeType),
+				Msg: fmt.Sprintf("failed to map attribute type %v to []string",
+					attr.AttributeType),
 				FieldName: field.Type().Name(),
 				SrcName:   attr.AttributeName,
 				SrcType:   attr.AttributeType,
@@ -180,7 +262,8 @@ func bindAttrToIntSliceField(attrs []CiAttribute, field reflect.Value) (err erro
 		number, err := strconv.Atoi(trimmedValue)
 		if err != nil {
 			return &BindError{
-				Msg:       fmt.Sprintf("failed convert attribute value \"%v\" to []int: %v", trimmedValue, err.Error()),
+				Msg: fmt.Sprintf("failed convert attribute value \"%v\" to []int: %v",
+					trimmedValue, err.Error()),
 				FieldName: field.Type().Name(),
 				SrcName:   attr.AttributeName,
 				SrcType:   attr.AttributeType,
